https://screenrec.com/share/HtGOEmaWei
I was able to loop all the data from the other sheets on to the 1st sheet, and I was only able to loop the headers successfully to all sheets. 
I took a few screenshots, but the list is too long, so I took a video.
On a side note: I had a scheduled tutoring session where I was planning to get clarification/help on looping the HW correctly, but the tutor had technical difficulties, and he was not able to help me one on one. That was very disappointing and frustrating with the program. Due to a simple technical issue, I was not able to get the help that I paid for, and the fact that there were not attempts to find workarounds was even more disappointing.